Update to React Filter
