+++ index.html
@@ -174,7 +174,7 @@
         <div class="food-filter-container">
           <form id="food-name-filter">
             <label for="name-filter">Search By Name</label>
-            <input id="food-filter-input">
+            <div id="food-filter-input"></div>
           </form>
         </div>
         <div>
