+++ lib/index.js
@@ -1,4 +1,8 @@
+import ReactDOM from 'react-dom';
+import React, { Component } from "react";
+import Filter from './Components/Filter.js';
 var foodsRequests = require('./requests/foods')
+
 var foodsDiary = require('./requests/diaryFoods')
 var events = require('./eventListeners/eventListeners')
 var fileName = location.pathname.split('/').slice(-1)[0]
@@ -11,11 +15,14 @@ $(document).ready(() => {
   addFoodMealListener()
   addFoodListener()
   editFoodListener(fileName, originalFood)
-  filterFoodListener()
+  // filterFoodListener()
   deleteFoodMealListener()
   calorieSorter()
   events.deleteFoodListener(foodsRequests)
   events.setOriginalFoodListener(originalFood)
+  ReactDOM.render(
+    <Filter />,
+    document.getElementById('food-filter-input'))
 })
 
 const originalFood = {}
@@ -54,11 +61,11 @@ const editFoodListener = (fileName, originalFood) => {
   })
 }
 
-const filterFoodListener = () => {
-  $('#food-filter-input').on('keyup', event => {
-    filterFoods()
-  })
-}
+// const filterFoodListener = () => {
+//   $('#food-filter-input').on('keyup', event => {
+//     filterFoods()
+//   })
+// }
 
 const deleteFoodMealListener = () => {
   $('.meal-table').on('click', event => {
@@ -153,20 +160,20 @@ const setFoodData = (food) => {
   }
 }
 
-const filterFoods = () => {
-  var filter = $('#food-filter-input').val().toUpperCase()
-
-  let foodNameNodes = findNodesForFilter()
-
-  for(var i = 0; i < foodNameNodes.length; i++) {
-    var name = $(foodNameNodes[i])
-    if (name.text().toUpperCase().indexOf(filter) > -1) {
-      $(name.parent()[0]).css('display', '')
-    } else {
-      $(name.parent()[0]).css('display', 'none')
-    }
-  }
-}
+// const filterFoods = () => {
+//   var filter = $('#food-filter-input').val().toUpperCase()
+//
+//   let foodNameNodes = findNodesForFilter()
+//
+//   for(var i = 0; i < foodNameNodes.length; i++) {
+//     var name = $(foodNameNodes[i])
+//     if (name.text().toUpperCase().indexOf(filter) > -1) {
+//       $(name.parent()[0]).css('display', '')
+//     } else {
+//       $(name.parent()[0]).css('display', 'none')
+//     }
+//   }
+// }
 
 const newFoodSequence = () => {
   var nameInput = $('#food-name').val()
@@ -201,15 +208,15 @@ const clearChecked = () => {
   $('.food-item-checkbox:checkbox:checked').prop('checked', false )
 }
 
-const findNodesForFilter = () => {
-  var fileName = location.pathname.split('/').slice(-1)[0]
-  if (fileName === 'foods.html' || fileName === 'foods') {
-    var foodNameNodes = $('.food-item-name')
-  } else {
-    var foodNameNodes = $('div.foods-diary-container .food-item-name')
-  }
-  return foodNameNodes
-}
+// const findNodesForFilter = () => {
+//   var fileName = location.pathname.split('/').slice(-1)[0]
+//   if (fileName === 'foods.html' || fileName === 'foods') {
+//     var foodNameNodes = $('.food-item-name')
+//   } else {
+//     var foodNameNodes = $('div.foods-diary-container .food-item-name')
+//   }
+//   return foodNameNodes
+// }
 
 const calorieSorter = () => {
   var originalOrder = []
