+++ lib/requests/diaryFoods.js
@@ -13,7 +13,7 @@ const mealsAPIFetch = (id, method, extension) => {
 const getDiaryFoods = () => {
   foodsRequests.foodsAPIFetch('', 'GET')
     .then(response => handleResponse(response))
-    .then(foods => getEachDiaryFood(foods))
+    .then(foods => renderDiaryFood(foods))
     .catch(error => console.error({ error }))
 }
 
@@ -68,9 +68,9 @@ const handleResponse = (response) => {
     })
 }
 
-const getEachDiaryFood = (foods) => {
-  renderDiaryFood(foods)
-}
+// const getEachDiaryFood = (foods) => {
+//   renderDiaryFood(foods)
+// }
 
 const renderDiaryFood = (foods) => {
   ReactDOM.render(
